refactor(product-client): extract frame-pipeline lifecycle to keep the hook under the line cap (repo-shape)
The rung-4 pipeline wiring pushed use-transcript-stick-to-bottom.ts back
over the 400-line soft cap after inheriting the rung-3 splits. Move the
single-writer registration, the tab/window-resume glue, and disposal into
use-transcript-frame-pipeline-lifecycle.ts (ownership-correct: the
pipeline's lifecycle, not the pin engine's). Hook is 398 lines;
report_frontend_structure.py: 0 violations.
